LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/transformations/2d.transformation.setTransform.nonfinite.worker-expected.txt


PASS setTransform() with Infinity/NaN is ignored 
